Working exemaple with binary threshhold
added gut thickness
using using first index of perpendicular vector line if this index is greater than 30. this will avoid too big values (if first peak is not detected). this values should be small, so find peaks should filter them out and we will have better data in the end